move mempool into server
