--- pkg/core/sdk/signer_middleware.go
@@ -1,1 +1,0 @@
-package sdk
